+++ packages/orq-rc/src/orq_ai_sdk/models/filegetop.py
@@ -59,5 +59,5 @@ class FileGetResponseBody(BaseModel):
     workspace_id: str
     r"""The id of the resource"""
 
-    created: Optional[datetime] = parse_datetime("2025-08-14T13:56:24.891Z")
+    created: Optional[datetime] = parse_datetime("2025-08-15T06:01:03.134Z")
     r"""The date and time the resource was created"""