+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -125,7 +125,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01K2MDWC2GHNP4310DVPRP731Q"
+        "01K2P52NGXQ1CB6P6RDGERRHJW"
     )
     r"""The id of the resource"""